// src/app/api/generated/model/clasificar-adjunto-request.ts
/**
 * AKINE API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Reclasificacion de un adjunto. El contenido no se puede cambiar
 */
export interface ClasificarAdjuntoRequest { 
    /**
     * Nueva clasificacion, o null para no cambiarla
     */
    categoria?: ClasificarAdjuntoRequestCategoriaEnum;
    /**
     * Nuevo titulo, o null para no cambiarlo. Cadena vacia lo borra
     */
    titulo?: string;
}
export enum ClasificarAdjuntoRequestCategoriaEnum {
    DOCUMENTO_IDENTIDAD = 'DOCUMENTO_IDENTIDAD',
    CREDENCIAL_COBERTURA = 'CREDENCIAL_COBERTURA',
    CONSENTIMIENTO = 'CONSENTIMIENTO',
    AUTORIZACION = 'AUTORIZACION',
    COMPROBANTE = 'COMPROBANTE',
    OTRO = 'OTRO'
};